refactor(iota-types): gate messages_consensus out of wasm; inline metric stubs
Two simplifications driven by the reference commit's structure:

- messages_consensus is node-only consensus machinery (fastcrypto-tbls /
  tonic). Gate the whole module out on wasm instead of carving out the
  DKG subset: messages_consensus.rs reverts to byte-identical develop and
  messages_consensus_dkg.rs is removed. The only client types it exposed
  (CancelledTransaction, ConsensusCommitPrologueV1,
  ConsensusDeterminedVersionAssignments) are re-exports from
  iota_sdk_types, so transaction.rs now imports them from the source. The
  node-only messages_grpc module (carries SignedAuthorityCapabilitiesV1)
  and the BcsSignable impl for VersionedMisbehaviorReport are gated too.

- metrics: drop the separate wasm_metrics.rs file and gate in place. The
  native items get a #[cfg(not(wasm32))] attribute each (no
  re-indentation) and the no-op stubs are appended behind
  #[cfg(wasm32)], matching how the reference handled metrics. Execution
  still needs LimitsMetrics / BytecodeVerifierMetrics on wasm, so the
  stubs stay — just in one file.

wasm_multiaddr.rs remains the only added file (as in the reference): the
Multiaddr type appears in on-chain validator metadata the wasm build reads.
